Display price ranges on different screen
Users may have a hard time dealing with random prices in a vacuum. It's
useful for them to know where it fits on the range of possible prices
for an item.

This allows the player to click a button to be shown a different screen
with the price ranges before returning to the main game screen.

This UX is inspired by the web version of the game we based this off of.

https://trello.com/c/q92hjQYN